refactor runtime props type
