+++ packages/runtime/src/traits/core/arrayState.tsx
@@ -1,6 +1,7 @@
 import { createTrait } from '@meta-ui/core';
 import { Static, Type } from '@sinclair/typebox';
-import { TraitImplementation } from '../../modules/registry';
+import { TraitImplementation } from 'src/types/RuntimeSchema';
+
 const HasInitializedMap = new Map<string, boolean>();
 
 type KeyValue = { key: string; value: unknown };
