+++ packages/runtime/src/traits/core/hidden.tsx
@@ -1,7 +1,7 @@
 import { CSSProperties } from 'react';
 import { createTrait } from '@meta-ui/core';
 import { Static, Type } from '@sinclair/typebox';
-import { TraitImplementation } from '../../modules/registry';
+import { TraitImplementation } from 'src/types/RuntimeSchema';
 
 const useHiddenTrait: TraitImplementation<Static<typeof PropsSchema>> = ({ hidden }) => {
   const style: CSSProperties = {};
