+++ packages/runtime/src/types/RuntimeSchema.ts
@@ -13,16 +13,27 @@ export type MetaUIModules = {
   apiService: ApiService;
 };
 
+export type ComponentWrapperType = React.FC<{ id: string }>;
+
+export type ComponentParamsFromApp = {
+  onLayoutChange?: (id: string, layout: any) => void;
+  componentWrapper?: ComponentWrapperType;
+};
+
 export type AppProps = {
   options: Application;
   mModules: MetaUIModules;
-  onLayoutChange?: (id: string, layout: any) => void;
-  componentWrapper?: ComponentWrapperType;
   debugStore?: boolean;
   debugEvent?: boolean;
-};
+} & ComponentParamsFromApp;
 
-export type ComponentWrapperType = React.FC<{ id: string }>;
+export type ImplWrapperProps = {
+  component: ApplicationComponent;
+  slotsMap: SlotsMap | undefined;
+  targetSlot: { id: string; slot: string } | null;
+  mModules: MetaUIModules;
+  app?: RuntimeApplication;
+} & ComponentParamsFromApp;
 
 export type SlotComponentMap = Map<string, SlotsMap>;
 export type SlotsMap = Map<
@@ -32,3 +43,36 @@ export type SlotsMap = Map<
     id: string;
   }>
 >;
+
+export type CallbackMap = Record<string, () => void>;
+
+export type SubscribeMethods = <U>(map: {
+  [K in keyof U]: (parameters: U[K]) => void;
+}) => void;
+export type MergeState = (partialState: any) => void;
+
+type RuntimeFunctions = {
+  mergeState: MergeState;
+  subscribeMethods: SubscribeMethods;
+};
+
+export type ComponentMergedProps = ImplWrapperProps &
+  TraitResult['props'] &
+  RuntimeFunctions;
+
+export type TraitResult = {
+  props: {
+    data?: unknown;
+    style?: Record<string, any>;
+    callbackMap?: CallbackMap;
+    effects?: Array<() => void>;
+  } | null;
+};
+
+export type TraitImplementation<T = any> = (
+  props: T &
+    RuntimeFunctions & {
+      componentId: string;
+      mModules: MetaUIModules;
+    }
+) => TraitResult;
